# UniversityApp/swagger.yaml
openapi: 3.0.1
info:
  title: University API
  description: This is a university api documentation
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /course/delete:
    get:
      operationId: deleteCourse
      parameters:
      - name: courseId
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: string
...
  /student/delete:
    get:
      operationId: deleteStudent
      parameters:
      - name: studentId
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: string
...
  /course/saveCourse:
    post:
      operationId: saveCourse
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/Course'
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: string
...
  /api/student/searchByFirstName/{firstName}:
    get:
      tags:
      - Student
      summary: Finds students by first name
      operationId: searchByFirstName
      parameters:
      - name: firstName
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Student'
...
  /api/lecture/search/{dateTime}:
    get:
      tags:
      - Lecture
      summary: Finds lectures by date
      operationId: searchByName
      parameters:
      - name: dateTime
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Lecture'
...
  /api/course/search/{name}:
    get:
      tags:
      - Course
      summary: Finds courses by name
      operationId: searchByName_1
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Course'
  /student/searchByFirstName:
    get:
      operationId: searchByFirstName_1
      parameters:
      - name: firstName
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: string
...
  /teacher/searchByLastName:
    get:
      operationId: searchByLastName
      parameters:
      - name: lastName
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: string
...
  /course/formForUpdate:
    get:
      operationId: updateCourse
      parameters:
      - name: courseId
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: string
...
  /student/formForUpdate:
    get:
      operationId: updateStudent
      parameters:
      - name: studentId
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: string
...
  /student/deleteFromGroup:
    get:
      operationId: deleteFromGroup
      parameters:
      - name: studentId
        in: query
        required: true
        schema:
          type: integer
          format: int32
      - name: groupId
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: string
  /teacher/searchByFirstName:
    get:
      operationId: searchByFirstName_2
      parameters:
      - name: firstName
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: string
...
  /student/saveStudent:
    post:
      operationId: saveStudent
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/Student'
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: string
...
  /api/teacher/searchByLastName/{lastName}:
    get:
      tags:
      - Teacher
      summary: Finds teachers by last name
      operationId: searchByLastName_1
      parameters:
      - name: lastName
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Teacher'
...
  /api/group/search/{name}:
    get:
      tags:
      - Group
      summary: Finds groups by name
      operationId: searchByName_2
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Group'
  /api/faculty/search/{name}:
    get:
      tags:
      - Faculty
      summary: Finds faculties by name
      operationId: searchByName_3
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Faculty'
...
  /api/student/searchByLastName/{lastName}:
    get:
      tags:
      - Student
      summary: Finds students by last name
      operationId: searchByLastName_2
      parameters:
      - name: lastName
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Student'
...
  /api/student/{studentId}/deleteFromGroup:
    delete:
      tags:
      - Student
      summary: Deletes student from its faculty
      operationId: deleteFromGroup_1
      parameters:
      - name: studentId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: string
...
  /api/teacher/searchByFirstName/{firstName}:
    get:
      tags:
      - Teacher
      summary: Finds teachers by first name
      operationId: searchByFirstName_3
      parameters:
      - name: firstName
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Teacher'
...
  /student/searchByLastName:
    get:
      operationId: searchByLastName_3
      parameters:
      - name: lastName
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: string
...
components:
  schemas:
    Audience:
      required:
      - audienceNumber
      - seats
      type: object
      properties:
        id:
          type: integer
          format: int32
        audienceNumber:
          maximum: 999
          minimum: 1
          type: integer
          format: int32
        seats:
          maximum: 200
          minimum: 10
          type: integer
          format: int32
        hasProjector:
          type: boolean
    Faculty:
      required:
      - description
      - name
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          maxLength: 2147483647
          minLength: 3
          type: string
        description:
          maxLength: 2147483647
          minLength: 10
          type: string
    Group:
      required:
      - name
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          maxLength: 2147483647
          minLength: 2
          type: string
...
    Course:
      required:
      - description
      - name
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          maxLength: 2147483647
          minLength: 3
          type: string
        description:
          maxLength: 2147483647
          minLength: 10
          type: string
    Student:
      required:
      - email
      - firstName
      - lastName
      type: object
      properties:
        id:
          type: integer
          format: int32
        firstName:
          maxLength: 2147483647
          minLength: 2
          type: string
        lastName:
          maxLength: 2147483647
          minLength: 2
          type: string
        email:
          type: string
    Lecture:
      type: object
      properties:
        id:
          type: integer
          format: int32
        course:
          $ref: '#/components/schemas/Course'
        audience:
          $ref: '#/components/schemas/Audience'
        teacher:
          $ref: '#/components/schemas/Teacher'
        dateTime:
          type: string
          format: date-time
        groups:
          type: array
          items:
            $ref: '#/components/schemas/Group'
    Teacher:
      required:
      - firstName
      - lastName
      type: object
      properties:
        id:
          type: integer
          format: int32
        firstName:
          maxLength: 2147483647
          minLength: 2
          type: string
        lastName:
          maxLength: 2147483647
          minLength: 2
          type: string
...